Prevent gesture processing from offscreen trackers
The tracker position isn't updated if the area is below the threshold.
This means a tracker not detected at all (i.e. if it is off screen)
will retain its previous position.

Any gestures that matched its position directly before it went offscreen
would then continue to match for the entire time it was offscreen.

By setting a "detected" flag every frame,
the gesture processor can immediately recognise any offscreen tracker
as not matching the gesture.