Add rough code for displaying webcam frame
A single frame of video is pulled from an OpenCV video capture device
then uploaded as an OpenGL texture.

The fragment shader flips the image vertically,
converts each pixel to HSV and back again (to prove it works),
and the image is displayed on screen.

This is a good starting point for using compute shaders to calculate inputs.