Update code to enable  minimal proof of concept
A source image is uploaded as a texture,
converted to HSV and stored in a texture that's blank at the start,
then converted back to RGB and stored in another blank texture.

All three are displayed by OpenCV at the end.
The HSV texture looks correctly incorrect due to displaying HSV as BGR.